requirements.txt
alabaster==0.7.12
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
bcrypt==3.1.7
bitstring==3.1.6
blist==1.3.6
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
cvxpy==1.1.7
cycler==0.10.0
Cython==0.29.13
deap==1.3.0
decorator==4.4.0
docopt==0.6.2
docutils==0.15.2
ecdsa==0.13.2
ecos==2.0.7.post1
future==0.17.1
gluoncv==0.6.0
graphviz==0.8.4
idna==2.8
imagesize==1.1.0
importlib-metadata==2.0.0
ipaddress==1.0.22
Jinja2==2.10.1
joblib==0.13.2
kiwisolver==1.1.0
liac-arff==2.4.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mock==3.0.5
more-itertools==7.2.0
mpi4py==3.0.2
mpmath==1.1.0
munch==2.5.0
netaddr==0.7.19
netifaces==0.10.9
nose==1.3.7
numpy==1.17.3
osqp==0.6.1
packaging==20.4
pandas==0.25.3
paramiko==2.6.0
pathlib2==2.3.4
pbr==5.4.3
Pillow==7.0.0
Pillow-SIMD==6.0.0.post0
pluggy==0.13.0
portalocker==1.5.2
pretrainedmodels==0.7.4
prettytable==0.7.2
protobuf==3.10.0
psutil==5.6.3
ptflops==0.6.6
py==1.8.0
py-expression-eval==0.3.9
pyasn1==0.4.7
pycparser==2.19
pycrypto==2.6.1
Pygments==2.4.2
PyNaCl==1.3.0
pyparsing==2.4.2
pytest==5.1.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
scandir==1.10.0
scikit-learn==0.23.2
scipy==1.3.1
scs==2.1.2
setuptools-scm==3.3.3
six==1.12.0
sklearn==0.0
snowballstemmer==1.9.1
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.1.2
tabulate==0.8.3
tensorboard-logger==0.1.0
tensorboardX==2.1
threadpoolctl==2.1.0
timm==0.4.12
torch==1.8.0
torchvision==0.9.0a0
tqdm==4.43.0
typing-extensions==3.7.4.3
ujson==1.35
urllib3==1.25.3
virtualenv==16.7.5
wcwidth==0.2.5
xlrd==1.2.0
zipp==0.6.0
